Displayed all the possible shiny pokemon using API and URL calls
Added Axios to the project for easy JSON data stringification and so that the project can be supported on a wider range of browsers.

Created a pokemon object to be used in the pokemon-list function to get the pokemon's name, id, and shiny sprite. The name and id come from PokeAPI and the shiny sprites come from this link: https://db.pokemongohub.net/tools/shiny-checklist

Created an array of pokemon ids that can be shiny in pokemon go and used this array to determine which pokemon would be fetched from the api. I then displayed all of these pokemon in their own elements on the shiny-collection page.

Some of the pokemon sprite URLs require more data injected into them to give a result. This is what I'll be working on next along with general styling changes to make the page feel and look better.