Added a filter function to the website.
Users can now filter pokemon by their generations. This keeps everything a bit more organized and makes it easier for users to find the pokemon they want to add to their collection. I also added a fallback url for the shiny sprites so that most (not all) of the shiny sprites appear in the list.